# faster implementation of erdos_renyi