Master : 
``` julia
julia> @benchmark erdos_renyi(500, 0.5)
BenchmarkTools.Trial: 115 samples with 1 evaluation.
 Range (min … max):  31.674 ms … 102.297 ms  ┊ GC (min … max): 0.00% … 14.32%
 Time  (median):     39.373 ms               ┊ GC (median):    0.00%
 Time  (mean ± σ):   43.619 ms ±  13.058 ms  ┊ GC (mean ± σ):  1.75% ±  5.16%

     ▇▂█▂▆▁▂ ▄ 
  ▄█████████▃█▆▆█▁▆▄▃▆▃▁▄▁▁▁▁▁▁▁▁▁▁▁▁▁▁▃▃▁▁▃▁▁▁▁▁▁▄▁▁▁▄▁▃▁▁▁▁▃ ▃
  31.7 ms         Histogram: frequency by time         90.9 ms <

 Memory estimate: 3.69 MiB, allocs estimate: 2502.

julia> @benchmark erdos_renyi(5000, 0.5)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 17.436 s (1.51% GC) to evaluate,
 with a memory estimate of 263.33 MiB, over 35003 allocations.

julia> @benchmark erdos_renyi(500, is_directed = true, 0.5)
BenchmarkTools.Trial: 82 samples with 1 evaluation.
 Range (min … max):  58.535 ms … 67.722 ms  ┊ GC (min … max): 0.00% … 5.64%
 Time  (median):     61.109 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   61.466 ms ±  1.972 ms  ┊ GC (mean ± σ):  0.98% ± 2.18%

         █▂▂  ▅ ▅  █▂▂    ▂      ▂
  ████▁▁███████▅█▅▁████▁██████▅▅▅█▅▁▅▁█▁▁▅█▅▅▅▁▅▅▁▁▁▁▁▅▁▁▁▁▁▅ ▁
  58.5 ms         Histogram: frequency by time        66.9 ms <

 Memory estimate: 7.38 MiB, allocs estimate: 5003.

julia> @benchmark erdos_renyi(5000, is_directed = true, 0.5)
BenchmarkTools.Trial: 1 sample with 1 evaluation.
 Single result which took 37.086 s (0.62% GC) to evaluate,
 with a memory estimate of 526.66 MiB, over 70005 allocations.
```
This commit :
``` julia
julia> @benchmark erdos_renyi(500, 0.5)
BenchmarkTools.Trial: 847 samples with 1 evaluation.
 Range (min … max):  3.717 ms … 12.779 ms  ┊ GC (min … max): 0.00% … 55.33%
 Time  (median):     5.758 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   5.883 ms ±  1.727 ms  ┊ GC (mean ± σ):  8.63% ± 15.70%

  ▂             █▆▁▁        
  █▆▄▃▃▃▃▃▃▃▄▅▆▆████▄▃▃▃▁▁▂▁▂▂▂▁▂▂▂▂▂▁▂▁▁▁▁▂▁▃▃▂▂▃▂▃▂▃▂▃▂▂▂▂ ▃
  3.72 ms        Histogram: frequency by time        11.7 ms <

 Memory estimate: 3.69 MiB, allocs estimate: 2502.

julia> @benchmark erdos_renyi(5000, 0.5)
BenchmarkTools.Trial: 7 samples with 1 evaluation.
 Range (min … max):  519.211 ms …    1.668 s  ┊ GC (min … max):  6.73% … 67.66%
 Time  (median):     586.219 ms               ┊ GC (median):    11.76%
 Time  (mean ± σ):   752.332 ms ± 411.881 ms  ┊ GC (mean ± σ):  29.61% ± 23.51%

  █▁ ▁    ▁  ▁                                                ▁  
  ██▁█▁▁▁▁█▁▁█▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█ ▁
  519 ms           Histogram: frequency by time          1.67 s <

 Memory estimate: 263.33 MiB, allocs estimate: 35003.

julia> @benchmark erdos_renyi(500, is_directed = true, 0.5)
BenchmarkTools.Trial: 576 samples with 1 evaluation.
 Range (min … max):  7.551 ms … 13.974 ms  ┊ GC (min … max): 0.00% … 23.43%
 Time  (median):     8.205 ms              ┊ GC (median):    0.00%
 Time  (mean ± σ):   8.665 ms ±  1.101 ms  ┊ GC (mean ± σ):  5.53% ±  9.83%

  ▃▃▂▂▂▂ ▅█▇▄▃▁▂    ▁                    ▁  ▁▁▁▁ ▁
  ███████████████▇▇▇█▅▆▄▁▄▆▅▆▁▁▁▇▁▁▁▁▄▄▁▇███████▄█▆▅▇▇▆▄▅▄▁▄ ▇
  7.55 ms      Histogram: log(frequency) by time     11.7 ms <

 Memory estimate: 7.38 MiB, allocs estimate: 5003.

julia> @benchmark erdos_renyi(5000, is_directed = true, 0.5)
BenchmarkTools.Trial: 4 samples with 1 evaluation.
 Range (min … max):  1.143 s …    1.453 s  ┊ GC (min … max):  0.00% … 22.82%
 Time  (median):     1.376 s               ┊ GC (median):    21.56%
 Time  (mean ± σ):   1.337 s ± 135.752 ms  ┊ GC (mean ± σ):  17.84% ± 11.62%

  █                                     █         █        █  
  █▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁▁█▁▁▁▁▁▁▁▁█ ▁
  1.14 s         Histogram: frequency by time         1.45 s <

 Memory estimate: 526.66 MiB, allocs estimate: 70005.
```