@etiennedeg 2022-06-24
Oh ok, I see the point of using the binomial distribution. The biggest problem here is the use of `add_edge!`, which is very inefficient. I will check your code to see if it gives improvement when I will have time.