@simonschoelly 2022-07-02
I mean, if we have some benchmarks, that show that this implementation is also faster for very small and very large `p` such as  `p = k/n` and `p = (n-k)/n`  (for some small integer `k`) then we could already merge this.